+++ state-chain/traits/src/mocks.rs
@@ -21,6 +21,7 @@ pub mod ensure_witnessed;
 pub mod epoch_info;
 pub mod eth_environment_provider;
 pub mod fee_payment;
+pub mod fetches_transfers_limit_provider;
 pub mod flip_burn_info;
 pub mod funding_info;
 pub mod ingress_egress_fee_handler;
